requirements.txt
fastapi
uvicorn[standard]
sqlmodel
asyncpg
databases
psycopg2
alembic
python-dotenv
twilio
pytest
httpx
gunicorn
openai